move snippet area into an own controller
