+++ src/Sulu/Bundle/SnippetBundle/Tests/Unit/Snippet/DefaultSnippetManagerTest.php
@@ -34,8 +34,8 @@ class DefaultSnippetManagerTest extends \PHPUnit_Framework_TestCase
                 'de' => 'Test DE',
             ],
         ],
-        'test.homepage' => [
-            'key' => 'test.homepage',
+        'test_homepage' => [
+            'key' => 'test_homepage',
             'template' => 'test',
             'title' => [
                 'en' => 'Test Homepage EN',
@@ -51,10 +51,10 @@ class DefaultSnippetManagerTest extends \PHPUnit_Framework_TestCase
             ['sulu_io', 'de', 'test', 'test', '123-123-123', false],
             ['sulu_io', 'de', 'test', 'test', '123-123-123', true, false],
             ['sulu_io', 'de', 'test', 'test', '123-123-123', false, false],
-            ['sulu_io', 'de', 'test', 'test.homepage', '123-123-123'],
-            ['sulu_io', 'de', 'test', 'test.homepage', '123-123-123', false],
-            ['sulu_io', 'de', 'test', 'test.homepage', '123-123-123', true, false],
-            ['sulu_io', 'de', 'test', 'test.homepage', '123-123-123', false, false],
+            ['sulu_io', 'de', 'test', 'test_homepage', '123-123-123'],
+            ['sulu_io', 'de', 'test', 'test_homepage', '123-123-123', false],
+            ['sulu_io', 'de', 'test', 'test_homepage', '123-123-123', true, false],
+            ['sulu_io', 'de', 'test', 'test_homepage', '123-123-123', false, false],
         ];
     }
